@@ -40,27 +40,29 @@
   "homepage": "https://github.com/signalwire/signalwire-node",
   "license": "MIT",
   "dependencies": {
-    "dotenv": "^4.0.0",
+    "dotenv": "^6.2.0",
     "faye-websocket": "^0.11.1",
-    "twilio": "^3.26.1",
+    "twilio": "^3.28.0",
     "uuid": "^3.3.2",
-    "winston": "^2.4.0"
+    "winston": "^3.2.1"
   },
   "devDependencies": {
-    "@types/dotenv": "^4.0.2",
-    "@types/jest": "^23.3.11",
-    "@types/node": "^9.4.6",
-    "@types/uuid": "^3.4.3",
-    "typescript": "^3.2.4",
-    "jest": "^23.6.0",
-    "ts-jest": "^23.10.5",
+    "@types/dotenv": "^6.1.0",
+    "@types/jest": "^24.0.6",
+    "@types/node": "^11.9.4",
+    "@types/uuid": "^3.4.4",
+    "jest": "^24.1.0",
+    "mockttp": "^0.12.6",
+    "ts-jest": "^24.0.0",
     "ts-loader": "^5.3.3",
     "tslint": "^5.12.1",
-    "mockttp": "^0.12.6",
-    "webpack": "^4.29.0",
-    "webpack-cli": "^3.2.1",
+    "typescript": "^3.3.3",
+    "graphql": "^0.12.0",
+    "webpack-cli": "^3.2.3",
     "webpack-dev-server": "^3.1.14",
-    "uglifyjs-webpack-plugin": "^2.1.1",
+    "acorn": "^6.0.0",
+    "babel-core": "^6.26.0",
+    "webpack": "^4.29.5",
     "babel-loader": "^7.1.5",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-env": "^1.7.0"
